requirements.txt
Django==3.1.2
pylint==2.6.0
django-taggit==1.3.0